# utr-api/spec/utr-api-spec.yaml
openapi: 3.0.3
info:
  title: UTR API
  description: Úszóverseny támogató rendszer API
  version: 0.3.0-alpha
servers:
  - url: 'https://utr.hu'
  - url: 'http://localhost:8080'
paths:
  /api/auth/login:
    post:
      summary: "Bejelentkezés"
      operationId: "authenticateUser"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JwtResponse"
      tags:
        - "Hitelesítés"
  /api/auth/users/:
    get:
      summary: "Összes felhasználó"
      operationId: "getAllUsers"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/UserPublicResponse"
      tags:
        - "Hitelesítés"
    put:
      summary: "Új felhasználó létrehozása"
      operationId: "createUser"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewUserRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
      tags:
        - "Hitelesítés"
  /api/auth/users/{userId}:
    get:
      summary: "Felhasználó publikus adatai"
      operationId: "getUser"
      parameters:
        - name: "userId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserPublicResponse"
      tags:
        - "Hitelesítés"
    delete:
      summary: "Felhasználó törlése"
      operationId: "deleteUser"
      parameters:
        - name: "userId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
      tags:
        - "Hitelesítés"
  /api/auth/users/{userId}/display-name:
    patch:
      summary: "Felhasználó nevének módosítása"
      operationId: "changeUserDisplayName"
      parameters:
        - name: "userId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "displayName"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
      tags:
        - "Hitelesítés"
  /api/auth/users/{userId}/password:
    patch:
      summary: "Felhasználó jelszavának módosítása"
      operationId: "changeUserPassword"
      parameters:
        - name: "userId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
      tags:
        - "Hitelesítés"
  /api/auth/users/{userId}/roles:
    patch:
      summary: "Felhasználó szerepköreinek módosítása"
      operationId: "changeUserRoles"
      parameters:
        - name: "userId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "role"
          in: "query"
          required: true
          schema:
            type: "array"
            items:
              type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
      tags:
        - "Hitelesítés"
...
components:
  schemas:
    LoginRequest:
      type: "object"
      properties:
        username:
          type: "string"
        password:
          type: "string"
    JwtResponse:
      type: "object"
      properties:
        token:
          type: "string"
        type:
          type: "string"
        id:
          type: "integer"
          format: "int64"
        username:
          type: "string"
        displayName:
          type: "string"
        roles:
          type: "array"
          items:
            type: "string"
...
    UserPublicResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        username:
          type: "string"
        displayName:
          type: "string"
        roles:
          type: "array"
          items:
            type: "string"
            enum:
              - "ROLE_ADMIN"
              - "ROLE_IDOROGZITO"
              - "ROLE_ALLITOBIRO"
              - "ROLE_SPEAKER"
    NewUserRequest:
      type: "object"
      properties:
        username:
          type: "string"
        displayName:
          type: "string"
        role:
          type: "array"
          items:
            type: "string"
        password:
          type: "string"
    ChangePasswordRequest:
      type: "object"
      properties:
        oldPassword:
          type: "string"
        newPassword:
          type: "string"
...
    MessageResponse:
      type: "object"
      properties:
        message:
          type: "string"
...